refactor: 테스트 파일에서 index.ts의 경로를 utils/index.ts로 변경 #1114
