+++ frontend/__tests__/useUserIdInput.test.ts
@@ -4,7 +4,7 @@ import { ERROR_MESSAGE } from '../src/common/constants/errorMessage';
 import { ID } from '../src/common/constants/id';
 import useUserIdInput from '../src/common/hooks/useUserIdInput';
 
-import { renderHook, act } from './index';
+import { renderHook, act } from './utils/index';
 
 const createFakeChangeEvent = (
   value: string,
