+++ frontend/__tests__/utils/index.ts
@@ -1,0 +1,2 @@
+export * from '@testing-library/react';
+export { render } from './customRender';
